Move code.google.com/p/codesearch into hound/codesearch to avoid collisions during go get.
