+++ src/hound/codesearch/index/regexp.go
