+++ src/hound/codesearch/index/regexp_test.go
