+++ src/hound/codesearch/index/write.go
@@ -14,7 +14,7 @@ import (
 	"syscall"
 	"unsafe"
 
-	"code.google.com/p/codesearch/sparse"
+	"hound/codesearch/sparse"
 )
 
 // Index writing.  See read.go for details of on-disk format.
