+++ src/hound/codesearch/lib/README.template
