+++ src/hound/codesearch/lib/buildall
