+++ src/hound/codesearch/regexp/copy.go
